chore: rename helper file localStorage to storage
