+++ src/pages/[integration].tsx
@@ -3,7 +3,7 @@ import React, { useEffect } from "react";
 
 import Integration from "@/components/Integration";
 import { fetchIntegrationPageStaticProps } from "@/lib/helpers/fetchers";
-import { storeSelectedIntegration } from "@/lib/helpers/localStorage";
+import { storeSelectedIntegration } from "@/lib/helpers/storage";
 import activeIntegrations from "@/lib/integrations/active";
 import { deserializeSchedule } from "@/lib/serialization/deserializers";
 import { IntegrationPageParams, RezervoIntegration } from "@/types/integration";
